Remove EDH leaderboard api methods & tests
